# API/authenticator/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentation of the API for Authenticator EFREI",
    "version" : "v1",
    "title" : "Swagger Authenticator EFREI",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "authenticator-efrei.azurewebsites.net",
  "basePath" : "/api",
  "schemes" : [ "http" ],
  "paths" : {
    "/api/users" : {
      "post" : {
        "summary" : "Register a new user based on user entity",
        "description" : "",
        "operationId" : "registerUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "User entity",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SignUpRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Login is already taken"
          }
        }
      }
    },
    "/api/users/login" : {
      "post" : {
        "summary" : "Login into the application",
        "description" : "",
        "operationId" : "authenticateUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Login entity",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/LoginRequestDTO"
          }
        }, {
          "name" : "url",
          "in" : "query",
          "description" : "Url of a website",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Url invalid or login invalid"
          }
        }
      }
    },
    "/api/users/pincode" : {
      "get" : {
        "summary" : "Get a pincode based on token",
        "description" : "",
        "operationId" : "getPincode",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "description" : "Token of a user",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Token invalid"
          }
        }
      }
    },
    "/api/users/token" : {
      "post" : {
        "summary" : "Create token based on login entity",
        "description" : "",
        "operationId" : "token",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Login entity",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/LoginRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/api/users/websites" : {
      "get" : {
        "summary" : "Get websites of a user based on token",
        "description" : "",
        "operationId" : "getWebsites",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "description" : "Token of a user",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Token invalid"
          }
        }
      }
    },
    "/api/users/websites/action-required" : {
      "get" : {
        "summary" : "Get websites which need action based on token",
        "description" : "",
        "operationId" : "getWebsitesActionRequired",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "description" : "Token of a user",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserWebsite"
              }
            }
          },
          "400" : {
            "description" : "Token invalid"
          }
        }
      }
    },
    "/api/websites" : {
      "post" : {
        "summary" : "Register a website based on website entity",
        "description" : "",
        "operationId" : "registerWebiste",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Website entity",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CreateWebsiteDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Token invalid or site already registered"
          }
        }
      }
    },
    "/api/websites/action-required" : {
      "post" : {
        "summary" : "Create a double verification based on url and token",
        "description" : "",
        "operationId" : "setActionRequired",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "description" : "Token of a user",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "url",
          "in" : "header",
          "description" : "Url of website",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Token invalid or url not register or user not register"
          }
        }
      }
    },
    "/api/websites/users" : {
      "get" : {
        "summary" : "Get users who are register to a specific website based on url and token",
        "description" : "",
        "operationId" : "getAllUser",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "description" : "Token of a user",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "url",
          "in" : "header",
          "description" : "Url of website",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/UserWebsite"
              }
            }
          },
          "400" : {
            "description" : "Token invalid or url not register or user isn't admin"
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "basicAuth" : {
      "type" : "basic"
    }
  },
  "definitions" : {
    "CreateWebsiteDTO" : {
      "type" : "object",
      "properties" : {
        "site" : {
          "type" : "string"
        },
        "managerToken" : {
          "type" : "string"
        }
      }
    },
    "LoginRequestDTO" : {
      "type" : "object",
      "properties" : {
        "usernameOrEmail" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        }
      }
    },
    "SignUpRequestDTO" : {
      "type" : "object",
      "properties" : {
        "username" : {
          "type" : "string",
          "minLength" : 3,
          "maxLength" : 40
        },
        "email" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 40
        },
        "password" : {
          "type" : "string",
          "minLength" : 6,
          "maxLength" : 50
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "username" : {
          "type" : "string",
          "minLength" : 3,
          "maxLength" : 40
        },
        "email" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 40
        },
        "password" : {
          "type" : "string",
          "minLength" : 6,
          "maxLength" : 100
        },
        "pincode" : {
          "type" : "string"
        },
        "websites" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/UserWebsite"
          }
        },
        "adminWebsites" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/Website"
          }
        }
      }
    },
    "UserWebsite" : {
      "type" : "object",
      "properties" : {
        "primaryKey" : {
          "$ref" : "#/definitions/UserWebsitesID"
        },
        "waiting" : {
          "type" : "boolean"
        },
        "user" : {
          "$ref" : "#/definitions/User"
        },
        "website" : {
          "$ref" : "#/definitions/Website"
        }
      }
    },
    "UserWebsitesID" : {
      "type" : "object",
      "properties" : {
        "user" : {
          "$ref" : "#/definitions/User"
        },
        "website" : {
          "$ref" : "#/definitions/Website"
        }
      }
    },
    "Website" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "url" : {
          "type" : "string"
        },
        "users" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/UserWebsite"
          }
        },
        "admin" : {
          "$ref" : "#/definitions/User"
        }
      }
    }
  }
}